ai/requirements.txt
# AI Model Training and Conversion Dependencies

# Core ML libraries
tensorflow>=2.10.0
tensorflowjs>=4.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Data handling
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Model evaluation
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0

# Optional: For GPU support (if available)
# tensorflow-gpu>=2.10.0

# Optional: For advanced data augmentation
# albumentations>=1.3.0

# Optional: For model optimization
# tensorflow-model-optimization>=0.7.0
